Prompt 10
===
I detected a bug while testing and added the test "EX-005b" to reproduce it. Add the test and fix the bug.
===
It detected the bug and tried to fix it but noticed after running the tests twice that the fix wasn't working. It eventually fixed it and checked that all test pass.